% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gliab.R
\name{calc_gliab}
\alias{calc_gliab}
\title{title}
\usage{
calc_gliab(obj, beta, mu, sigma)
}
\arguments{
\item{obj}{placeholder}

\item{beta}{placeholder}

\item{mu}{placeholder}

\item{sigma}{placeholder}
}
\value{
placeholder
}
\description{
Helper function to calculate genetic liabilities.
}
