% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_normalization_consts.R
\name{calc_normalization_consts}
\alias{calc_normalization_consts}
\title{Constants for normalization}
\usage{
calc_normalization_consts(MAF, causal)
}
\arguments{
\item{MAF}{List of Minor Ale Frequency.}

\item{causal}{List of causal SNP´s where causal SNP´s has the value 1.}
}
\value{
A list with the theoretical values of mu and sigma.
}
\description{
This function calculates the theoretical value of the constants mu and sigma, which are used to normalize the data.
}
