% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_calc.R
\name{cond_calc}
\alias{cond_calc}
\title{placeholder}
\usage{
cond_calc(i, covmatrix)
}
\arguments{
\item{i}{Index of person}

\item{covmatrix}{A co-variance matrix}
}
\value{
The conditional values of mu and sigma
}
\description{
This function calculates the conditional mu and sigma given a co-variance
matrix.
}
