% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmatrix.R
\name{covmatrix}
\alias{covmatrix}
\title{Co-variance matrix}
\usage{
covmatrix(h2, n_sib = 0)
}
\arguments{
\item{h2}{Heritability parameter.}

\item{sib}{Number of siblings.}
}
\value{
A co-variance matrix, created from the value of h2 and
the amount of sib.
}
\description{
This function creates the theoretical co-variance matrix from the
multivariate normal distribution which is used to model the liabilities.
}
