% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRds.R
\name{createRds}
\alias{createRds}
\title{Create .rds file.}
\usage{
createRds(path, nrow, ncol)
}
\arguments{
\item{path}{Path where to save and find the file.}

\item{nrow}{Amount of rows.}

\item{ncol}{Amount of columns.}
}
\value{
An rds object containing a list with a file-backed matrix, FAM and MAP information in a list.
}
\description{
This function is used create a file-backed matrices and save it as an .rds file with related information.
}
