% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{placeholder}
\usage{
gibbs_sampler(config, burn_in, cov_mat, prevalence)
}
\arguments{
\item{config}{List of configurations}

\item{burn_in}{}

\item{cov_mat}{Co-variance matrix}

\item{prevalence}{The likelihood of having the disease}
}
\value{
The mean of genetic liability from the given configaration
}
\description{
This function calculates mean of genetic liability given a configuration.
}
