% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.R
\name{GWAS}
\alias{GWAS}
\title{GWAS
Calculates a linear regression}
\usage{
GWAS(person, y, include = rows_along(person$genotypes))
}
\arguments{
\item{person}{A .rds file with an FBM.code256 and accompanying FAM and MAP.}

\item{y}{Vector of regressands to regress on}

\item{include}{Vector of rows to use in regression. Used with cross-validation. Default uses all rows.}
}
\value{
A data.frame with slopes of each regression, standard errors of each slope, t-scores associated with each slope and P-values of each slope.
}
\description{
GWAS
Calculates a linear regression
}
