% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFH.R
\name{LTFH}
\alias{LTFH}
\title{placeholder}
\usage{
LTFH(child, prevalence, h2)
}
\arguments{
\item{child}{Dataframe containing liabilities and status for the entire family}

\item{prevalence}{The likelihood of having the disease}

\item{h2}{Heredity}
}
\value{
placeholder
}
\description{
This function finds all the unique status configurations in the data and
calculates the mean posterior genetic liability for each of these unique
configurations using a gibbs-sampler. The calculated mean posterior genetic
liabilities are then matched to each person in the data-set.
}
