% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan_plot.R
\name{Manhattan_plot}
\alias{Manhattan_plot}
\title{Manhattan plot}
\usage{
Manhattan_plot(x)
}
\arguments{
\item{x}{A regression}
}
\value{
A Manhattan plot with each regressor on the x-axis and their -log10(P-values) on the y-axis.
}
\description{
Produces a Manhattan plot from a regression
}
