% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_cross_validation.R
\name{Prediction_cross_validation}
\alias{Prediction_cross_validation}
\title{Prediction Cross-validation}
\usage{
Prediction_cross_validation(
  person,
  k = 10,
  threshold = 0,
  disease,
  method = "GWAS",
  liabilities = person$FAM$Status
)
}
\arguments{
\item{person}{A .rds file with an FBM.code256 and accompanying FAM and MAP.}

\item{k}{Number of folds to be used in cross-validation. Default is 10.}

\item{threshold}{Vector of P-values to be used in thresholding. Default does not use thresholding.}

\item{disease}{List with properties of disease.}

\item{method}{Method to use for prediction. Possible methods are "GWAS", "GWAX", "LTFH". Default is "GWAS".}

\item{liabilities}{Vector of liabilities used for prediction with "LTFH" method. If not specified, uses "GWAS" method instead.}
}
\value{
A list with 3 values: a tibble with average and best scores for each threshold, the best score, and the best model.
}
\description{
Function to calculate predictive powers of different models
}
