% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_trunc.R
\name{rnorm_trunc}
\alias{rnorm_trunc}
\title{placeholder}
\usage{
rnorm_trunc(n, range, mu, sigma)
}
\arguments{
\item{n}{placeholder}

\item{range}{placeholder}

\item{mu}{placeholder}

\item{sigma}{placeholder}
}
\value{
The threshold value used to get the truncated distribution
}
\description{
Helper function to calculate threshold for truncated distribution
}
