% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyouSick-package.R
\docType{package}
\name{RyouSick-package}
\alias{RyouSick}
\alias{RyouSick-package}
\title{RyouSick: Simulation and Analysis of Genetic Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nicklas Stanislaus \email{dummy@dummy.com}

Authors:
\itemize{
  \item Markus Nielsen
  \item Christian Mains
}

}
\keyword{internal}
