% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_disease.R
\name{sim_disease}
\alias{sim_disease}
\title{Simulation of a disease}
\usage{
sim_disease(
  n_snp,
  prevalence,
  h2,
  causal = NULL,
  causal_n = round(n_snp/10),
  maf = NULL,
  maf_low = 0.01,
  maf_high = 0.49,
  seed = NULL
)
}
\arguments{
\item{n_snp}{Amount of SNPs}

\item{prevalence}{Prevalence of disease in simulated population}

\item{h2}{Herability parameter}

\item{causal}{List of predetermined causal SNPs. Leave empty to get random causal SNPs.}

\item{causal_n}{Amount of causal SNPs. Default value is 10 procent of total SNPs}

\item{maf}{List of predetermined Minor Ale Frequency}

\item{maf_low}{placeholder}

\item{maf_high}{placeholder}

\item{seed}{placeholder}
}
\value{
placeholder
}
\description{
This function is used to simulate a disease parameters.
}
