% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_genotypes_no_family.R
\name{sim_genotypes_no_family}
\alias{sim_genotypes_no_family}
\title{Description - Simulation with no family}
\usage{
sim_genotypes_no_family(n, disease, path, overwrite = T, n_blocks = 20)
}
\arguments{
\item{n}{placeholder}

\item{disease}{placeholder}

\item{path}{placeholder}

\item{overwrite}{placeholder}

\item{n_blocks}{placeholder}
}
\value{
placeholder
}
\description{
Description - Simulation with no family
}
