% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_genotypes_with_family.R
\name{sim_genotypes_with_family}
\alias{sim_genotypes_with_family}
\title{Description -}
\usage{
sim_genotypes_with_family(
  n,
  disease,
  path,
  n_sibs = NULL,
  overwrite = T,
  n_blocks = 20
)
}
\arguments{
\item{n}{placeholder}

\item{disease}{placeholder}

\item{path}{placeholder}

\item{n_sibs}{placeholder}

\item{overwrite}{placeholder}

\item{n_blocks}{placeholder}
}
\value{
placeholder
}
\description{
Simulation with family (Fixed and non fixed)
}
