% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyRds.R
\name{verifyRds}
\alias{verifyRds}
\title{Verification of .rds file.}
\usage{
verifyRds(path, overwrite, nrow, ncol)
}
\arguments{
\item{path}{Path where to save and find the file.}

\item{overwrite}{Boolean value.}

\item{nrow}{Amount of rows.}

\item{ncol}{Amount of columns.}
}
\description{
This function is used to verify if the path is valid. If the path is already in use,
it checks if the already exiting .rds file has the same dimensions and overwrites it.
If the dimensions are wrong it stops and returns a error message.
}
